Refactor out the request part in call_api and call_account_api
